services/telegram-bot/requirements.txt
# Arasul Telegram Bot Dependencies
# Python 3.11+

# Telegram Bot API (async, v21+)
python-telegram-bot==21.7

# HTTP client for backend communication
httpx==0.27.2

# Health check server
flask==3.0.0
werkzeug==3.0.1

# Utilities
python-dotenv==1.0.0

# ==========================================================================
# LLM Providers (Telegram Bot 2.0)
# ==========================================================================

# Claude API
anthropic==0.40.0

# Token counting
tiktoken==0.8.0

# ==========================================================================
# Voice Processing
# ==========================================================================

# Audio processing
pydub==0.25.1

# Whisper (local transcription) - optional, requires ffmpeg
# openai-whisper==20231117

# OpenAI API (for Whisper API)
openai==1.58.1

# ==========================================================================
# Database
# ==========================================================================

# PostgreSQL async driver
asyncpg==0.30.0

# ==========================================================================
# Encryption
# ==========================================================================

# Cryptography for AES-256-GCM
cryptography==44.0.0
